PiServer/Código/requirements.txt
flask
mcp
mysql-connector-python
